Move roots reconciliation function to consensus.rs
