+++ core/src/replay_stage.rs
@@ -3,7 +3,7 @@
 use crate::{
     cluster_info::ClusterInfo,
     commitment::{AggregateCommitmentService, BlockCommitmentCache, CommitmentAggregationData},
-    consensus::{StakeLockout, Tower},
+    consensus::{reconcile_blockstore_roots_with_tower, StakeLockout, Tower},
     poh_recorder::PohRecorder,
     result::Result,
     rewards_recorder_service::RewardsRecorderSender,
@@ -12,7 +12,6 @@ use crate::{
 use solana_ledger::{
     bank_forks::BankForks,
     blockstore::Blockstore,
-    blockstore_db,
     blockstore_processor::{
         self, BlockstoreProcessorError, ConfirmationProgress, ConfirmationTiming,
         TransactionStatusSender,
@@ -219,7 +218,7 @@ impl ReplayStage {
         };
 
         // Tower gets written to before blockstore, so roots may be missing from blocktstore
-        Self::reconcile_blockstore_roots_with_tower(&tower, &blockstore)
+        reconcile_blockstore_roots_with_tower(&tower, &blockstore)
             .expect("failed to reconcile blockstore roots with tower");
 
         // Start the replay stage loop
@@ -477,24 +476,6 @@ impl ReplayStage {
             },
             root_bank_receiver,
         )
-    }
-
-    fn reconcile_blockstore_roots_with_tower(
-        tower: &Tower,
-        blockstore: &Blockstore,
-    ) -> blockstore_db::Result<()> {
-        if let Some(troot) = tower.root() {
-            let broot = blockstore.last_root();
-            if broot < troot {
-                let new_roots: Vec<_> = blockstore
-                    .slot_meta_iterator(broot + 1)?
-                    .map(|(slot, _)| slot)
-                    .take_while(|slot| *slot <= troot)
-                    .collect();
-                return blockstore.set_roots(&new_roots);
-            }
-        }
-        Ok(())
     }
 
     fn log_leader_change(
@@ -1151,7 +1132,7 @@ pub(crate) mod tests {
         iter,
         sync::{Arc, RwLock},
     };
-    use trees::{tr, Tree};
+    use trees::tr;
 
     struct ForkInfo {
         leader: usize,
@@ -2039,74 +2020,5 @@ pub(crate) mod tests {
                 .fork_weight;
             assert!(second >= first);
         }
-    }
-
-    fn build_tree(n: u64) -> Tree<u64> {
-        _build_tree(n - 1, n - 1)
-    }
-    fn _build_tree(n: u64, i: u64) -> Tree<u64> {
-        if i > 0 {
-            tr(n - i) / _build_tree(n, i - 1)
-        } else {
-            tr(n)
-        }
-    }
-
-    #[test]
-    fn test_reconcile_blockstore_roots_with_tower() {
-        let blockstore_path = get_tmp_ledger_path!();
-        {
-            let blockstore = Blockstore::open(&blockstore_path).unwrap();
-            let node_keypair = Keypair::new();
-            let vote_keypair = Keypair::new();
-            let stake_keypair = Keypair::new();
-            let node_pubkey = node_keypair.pubkey();
-            let mut keypairs = HashMap::new();
-            keypairs.insert(
-                node_pubkey,
-                ValidatorVoteKeypairs::new(node_keypair, vote_keypair, stake_keypair),
-            );
-
-            let (bank_forks, mut progress) = initialize_state(&keypairs);
-            let bank_forks = Arc::new(RwLock::new(bank_forks));
-            let mut tower = Tower::new_with_key(&node_pubkey);
-
-            // Create the tree of banks in a BankForks object
-            let forks = build_tree(42);
-
-            let mut voting_simulator = VoteSimulator::new(&forks);
-            let mut cluster_votes: HashMap<Pubkey, Vec<Slot>> = HashMap::new();
-            let votes: Vec<Slot> = (0..42).collect();
-
-            for slot in &votes {
-                let slot = *slot;
-                let (shreds, _) = make_slot_entries(slot + 1, slot, 1);
-                blockstore.insert_shreds(shreds, None, false).unwrap();
-                assert_eq!(
-                    voting_simulator.simulate_vote(
-                        slot,
-                        &bank_forks,
-                        &mut cluster_votes,
-                        &keypairs,
-                        keypairs.get(&node_pubkey).unwrap(),
-                        &mut progress,
-                        &mut tower,
-                    ),
-                    VoteResult::Ok
-                );
-            }
-            let tower_root = tower.root().unwrap();
-            let set_roots: Vec<_> = (0..(tower_root / 2)).collect();
-            blockstore.set_roots(&set_roots).unwrap();
-            assert!(blockstore.last_root() < tower_root);
-            assert!(
-                ReplayStage::reconcile_blockstore_roots_with_tower(&tower, &blockstore).is_ok()
-            );
-            assert_eq!(blockstore.last_root(), tower.root().unwrap());
-            assert!(
-                ReplayStage::reconcile_blockstore_roots_with_tower(&tower, &blockstore).is_ok()
-            );
-        }
-        Blockstore::destroy(&blockstore_path).expect("Expected successful database destruction");
     }
 }
